+++ source/Halibut.Tests/Halibut.Tests.csproj
@@ -54,9 +54,5 @@
   <ItemGroup>
     <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Util\TcpUtils" />
-  </ItemGroup>
-
+  
 </Project>
