+++ source/Halibut.Tests/Support/BackwardsCompatibility/HalibutTestBinaryRunner.cs
@@ -29,7 +29,6 @@ namespace Halibut.Tests.Support.BackwardsCompatibility
             this.serviceCertAndThumbprint = serviceCertAndThumbprint;
             this.version = version;
         }
-
         public HalibutTestBinaryRunner(ServiceConnectionType serviceConnectionType, int? clientServicePort, CertAndThumbprint clientCertAndThumbprint, CertAndThumbprint serviceCertAndThumbprint, string version) :
             this(serviceConnectionType, clientCertAndThumbprint, serviceCertAndThumbprint, version)
         {
@@ -41,32 +40,7 @@ namespace Halibut.Tests.Support.BackwardsCompatibility
         {
             this.webSocketServiceEndpointUri = webSocketServiceEndpointUri;
         }
-
-        string BinaryDir(string version)
-        {
-            var onDiskVersion = version.Replace(".", "_");
-            var assemblyDir = new DirectoryInfo(Path.GetDirectoryName(typeof(HalibutTestBinaryRunner).Assembly.Location)!);
-            var upAt = assemblyDir.Parent.Parent.Parent.Parent;
-            var projectName = $"Halibut.TestUtils.CompatBinary.v{onDiskVersion}";
-            var executable = Path.Combine(upAt.FullName, projectName, assemblyDir.Parent.Parent.Name, assemblyDir.Parent.Name, assemblyDir.Name, projectName);
-            executable = AddExeForWindows(executable);
-            if (!File.Exists(executable))
-            {
-                throw new Exception("Could not executable at path:\n" +
-                                    executable + "\n" +
-                                    $"Did you forget to update the csproj to depend on {projectName}\n" +
-                                    "If testing a previously untested version of Halibut a new project may be required.");
-            }
-
-            return executable;
-        }
-
-        string AddExeForWindows(string path)
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows)) return path + ".exe";
-            return path;
-        }
-
+        
         public async Task<RunningOldHalibutBinary> Run()
         {
             var envs = new Dictionary<string, string>();
@@ -123,7 +97,7 @@ namespace Halibut.Tests.Support.BackwardsCompatibility
                         if (s.Contains("RunningAndReady")) hasTentacleStarted.Set();
                     };
 
-                    ShellExecutor.ExecuteCommand(BinaryDir(version),
+                    ShellExecutor.ExecuteCommand(new HalibutTestBinaryPath().BinPath(version),
                         "",
                         tmp.FullPath,
                         processLogs,
